define RememberGen in python rather than slim!
Having the remembered generations hard-coded on the SLiM side is really
inflexible, especially when re-using the code for simulating different
scenarios. Here, we move this functionality into the Arguments class on the
python side instead. Because we can't simply set a dataclass field to a default
List, we make use of a `__post_init__` hook to set the list using a custom
method.